Make numeric-literal parsing more robust.
For cases where large non-fp numeric literals might end up triggering
coversion or rounding errors when stored as doubles when lexing.

This is a corner case, but it does trigger a case or two in the ECMA5
test suite (test262).

+++ src/qml/qml/parser/qqmljslexer.cpp
@@ -891,7 +891,7 @@ again:
 int Lexer::scanNumber(QChar ch)
 {
     if (ch != QLatin1Char('0')) {
-        double integer = ch.unicode() - '0';
+        quint64 integer = ch.unicode() - '0';
 
         QChar n = _char;
         const QChar *code = _codePtr;
